pyproject.toml
[project]
name = "tekne-admin-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.5",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "agno>=0.1.0",
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "ruamel-yaml>=0.19.0",
    "redis>=5.0.0",
]
